AGent/hi/sovereign-job-hunter/pyproject.toml
[project]
name = "sovereign-job-hunter"
version = "0.1.0"
description = "Local-first, multi-agent job application pipeline with evidence-based verification"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
  "google-genai>=1.9.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.0.1",
  "pypdf>=5.0.0",
  "qdrant-client>=1.14.0",
  "reportlab>=4.2.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
